Newest versions of the files
I'm about to create algorithms to handle the rest of the functionality. I have actually had to take a step back from work to put more time into this. It is proving challenging to get back on the horse. I will be including this repository in my portion of our paper due today so that she can take a look at our work. Testing will happen when we have functionality of each process. What I mean by that is when we have the signup page done, that will be tested. While that is being tested, I will be working on the image upload page or another point of functionality. Any problems that come up during the testing will be addressed by Samia. If she needs help resolving the issues, then we will work things out together over a Zoom meeting where we can share our screens. This way we can continue with the CI/CD portion of this process. We will continue developing while testing and repairing.