New Files
These are the PHP files created to run the site on XAMPP. The way to make it work is to put the files into the HTDocs folder. This should be under C:\xampp\htdocs. Then launch it and take the link from the browser and put it into Postman. I'm about to do this myself. 

I changed the PHP script for connecting to the server. It should work better. Modify as needed for your system.