Search JavaScript
When the search button is clicked, the function in this JavaScript file will be activated. I'll code in the HTML for the search button tomorrow. I took two days off from work to work on this and try to get the rest of the site up and working before the end of the year so we can spend January and February focusing on the Android and iOS functionality.