Newest Changes
I added in the search form, and the corresponding CSS for it. It still needs a little bit of work, as you can see by the way it currently looks. I cleaned up the PHP code to validate the inputs, basically making sure that there's actually a value in the field. This got rid of all but one of the errors that we were experiencing before. Furthermore, I posted a question to StackOverflow because I finally found a problem that nobody has asked a question about before. Basically, I broke it in a whole new way.