All Problems Solved
All of our errors and CSS woes have been solved. We are now ready to enter the testing portion. We will be removing the signup and sign in functionality, because there's no purpose for it right now. In the future, we may make that sign in popup the signup popup. There is no CSS for the results div right now, and I'm still debating on whether or not to load the results on the same page, or separate page(s). For now though, we have functionality. Testing should begin with Postman. I think that will be the best software we can use to test our HTTP requests. Let me know what you guys think in the group text.